Added support for lower resolution preview rendering (using two stage Render-to-FrameBufferObject / Texture rendering) Fixed rounding error with 3D fly camera.  Closing animationController should switch to Automatic

+++ Fragmentarium-Source/SyntopiaCore/Math/Matrix4.h
@@ -194,7 +194,7 @@ namespace SyntopiaCore {
 			scalar v[16];
 		};
 
-		typedef Matrix4<float> Matrix4f ;
+		typedef Matrix4<double> Matrix4f ;
 		typedef Matrix4<double> Matrix4d ;
 	}
 }
